Updated install script

+++ kubernetes/scripts/setup_kubespray.sh
@@ -10,10 +10,16 @@ mkdir -p "$KUBESPRAY_DATA_DIR"
 rm -rf "$KUBESPRAY_DATA_DIR/*"
 chmod 700 "$KUBESPRAY_DATA_DIR"
 cd "$KUBESPRAY_DATA_DIR"
-curl -Lo kubespray.tar.gz \
-    "https://github.com/kubernetes-sigs/kubespray/archive/refs/tags/v$KUBESPRAY_VERSION.tar.gz"
-tar -xzvf kubespray.tar.gz
-rm kubespray.tar.gz
-mv "kubespray-$KUBESPRAY_VERSION" kubespray
-python3 -m venv env
+if [ -d kubespray ]; then
+    curl -Lo kubespray.tar.gz \
+        "https://github.com/kubernetes-sigs/kubespray/archive/refs/tags/v$KUBESPRAY_VERSION.tar.gz"
+    tar -xzvf kubespray.tar.gz
+    rm kubespray.tar.gz
+    mv "kubespray-$KUBESPRAY_VERSION" kubespray
+fi
+sed -i 's|^minimal_master_memory_mb:.*|minimal_master_memory_mb: 0|g' kubespray/roles/kubernetes/preinstall/defaults/main.yml
+sed -i 's|^minimal_node_memory_mb:.*|minimal_node_memory_mb: 0|g' kubespray/roles/kubernetes/preinstall/defaults/main.yml
+if [ ! -f env/bin/pip3 ]; then
+    python3 -m venv env
+fi
 env/bin/pip3 install -r kubespray/requirements.txt
